Move RosterServiceStateMock
