+++ platform-sdk/consensus-roster/src/testFixtures/java/module-info.java
@@ -11,5 +11,7 @@ open module org.hiero.consensus.roster.test.fixtures {
     requires org.hiero.base.utility.test.fixtures;
     requires org.hiero.consensus.utility;
     requires org.junit.jupiter.params;
+    requires org.hiero.consensus.platformstate;
+    requires org.mockito;
     requires static transitive com.github.spotbugs.annotations;
 }
