+++ platform-sdk/consensus-roster/src/testFixtures/java/org/hiero/consensus/roster/test/fixtures/RosterServiceStateMock.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.test.fixtures.roster;
+package org.hiero.consensus.roster.test.fixtures;
 
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
