+++ platform-sdk/swirlds-platform-core/src/testFixtures/java/module-info.java
@@ -27,7 +27,6 @@ open module com.swirlds.platform.core.test.fixtures {
 
     exports com.swirlds.platform.test.fixtures;
     exports com.swirlds.platform.test.fixtures.config;
-    exports com.swirlds.platform.test.fixtures.roster;
     exports com.swirlds.platform.test.fixtures.simulated;
     exports com.swirlds.platform.test.fixtures.state;
 }
